+++ src/main/java/github/fintechecommerce/model/ExemptionMessages.java
@@ -1,13 +1,14 @@
 package github.fintechecommerce.model;
 
 public enum ExemptionMessages {
-    IllegalArgumentException, JpaSystemException;
+    IllegalArgumentException, JpaSystemException, NullPointerException;
 
     public String getCustomMessage() {
 
         return switch (this) {
             case IllegalArgumentException -> "small";
             case JpaSystemException -> "ErrorLog in database";
+            case NullPointerException -> "NullPointerException";
         };
     }
 }
