+++ src/main/java/github/fintechecommerce/utils/ResponseExceptionHandler.java
@@ -1,10 +1,9 @@
 package github.fintechecommerce.utils;
 
-import github.fintechecommerce.entity.ErrorLog;
+import github.fintechecommerce.model.ErrorLog;
 import github.fintechecommerce.model.ExemptionMessages;
 import github.fintechecommerce.model.ResponseModel;
 import org.hibernate.HibernateException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.orm.jpa.JpaSystemException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -19,24 +18,23 @@ import java.time.LocalDateTime;
 @ControllerAdvice
 public class ResponseExceptionHandler extends ResponseEntityExceptionHandler {
 
-
-    private ErrorLog errorLog;
-
-
     @ExceptionHandler(value = {IllegalArgumentException.class, IllegalStateException.class})
-    protected ResponseModel<String> handleConflict(RuntimeException ex, WebRequest request) {
-        String bodyOfResponse = ExemptionMessages.IllegalArgumentException.getCustomMessage();
-        return new ResponseModel<>(HttpStatus.CONFLICT.value(), HttpStatus.CONFLICT.getReasonPhrase(), LocalDateTime.now(), bodyOfResponse);
+    @ResponseStatus(value = HttpStatus.CONFLICT)
+    protected @ResponseBody ResponseModel<ErrorLog> handleConflict(Exception ex, WebRequest request) {
+
+        var error = new ErrorLog(ExemptionMessages.IllegalArgumentException.getCustomMessage(), ex.getMessage(), request.getDescription(false));
+
+        return new ResponseModel<>(HttpStatus.CONFLICT.value(), HttpStatus.CONFLICT.getReasonPhrase(), LocalDateTime.now(), error);
 
     }
 
     @ExceptionHandler(value = {JpaSystemException.class, HibernateException.class})
     @ResponseStatus(value = HttpStatus.INTERNAL_SERVER_ERROR)
-    protected @ResponseBody ResponseModel<String> handleJpaSystemException(JpaSystemException ex, WebRequest request) {
+    protected @ResponseBody ResponseModel<ErrorLog> handleJpaSystemException(Exception ex, WebRequest request) {
 
-var x = new ErrorLog(ex.getMessage(), ex.getCause().toString(), ex.getStackTrace().toString()));
-        return new ResponseModel<>(HttpStatus.INTERNAL_SERVER_ERROR.value(), HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase(), LocalDateTime.now());
+        var error = new ErrorLog(ExemptionMessages.IllegalArgumentException.getCustomMessage(), ex.getMessage(), request.getDescription(false));
 
+        return new ResponseModel<>(HttpStatus.CONFLICT.value(), HttpStatus.CONFLICT.getReasonPhrase(), LocalDateTime.now(), error);
     }
 
 }
