+++ src/main/resources/application.properties
@@ -1,30 +1,36 @@
 api_prefix=/api/${version}
-
 spring.application.name=${application_name}
 server.port=${server_port}
-
-spring.datasource.url= jdbc:sqlserver://localhost:${database_port};databaseName=${database_name}
-spring.datasource.username= ${database_username}
-spring.datasource.password= ${database_password}
-spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.SQLServer2012Dialect
+spring.datasource.url=jdbc:sqlserver://localhost:${database_port};databaseName=${database_name}
+spring.datasource.username=${database_username}
+spring.datasource.password=${database_password}
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.SQLServer2012Dialect
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
-
-
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
-
 #To Be Updated
 application_name=ProductApi
 version=v1
-database_name= test
-database_password= pass@word1
-database_username= sa
+database_name=test
+database_password=pass@word1
+database_username=sa
 server_port=8076
 database_port=1433
-
 spring.redis.port=6379
 spring.redis.password=redis
 spring.redis.host=localhost
 spring.cache.redis.time-to-live=10m
 spring.cache.redis.cache-null-values=false
+#Spring Boot actuator properties
+management.security.enabled=true
+management.security.roles=ADMIN
+security.basic.enabled=true
+security.user.name=admin
+security.user.passowrd=admin
+
+#Switch Off whitelabel error pages
+server.error.whitelabel.enabled=true
+
+
+spring.mvc.throw-exception-if-no-handler-found=true
